refactor(frontend): reorganise imports and update utils
