+++ frontend/dashboard/src/routes/DashboardPage.tsx
@@ -8,7 +8,7 @@ import {
 } from "@mui/material";
 import React from "react";
 import { Link } from "react-router-dom";
-import WorkflowsNavbar from "workflows-lib/lib/components/workflow/WorkflowsNavbar";
+import { WorkflowsNavbar } from "workflows-lib";
 
 const workflowsCard = (
   <React.Fragment>
