+++ frontend/workflows-lib/lib/main.ts
@@ -4,4 +4,7 @@ export { default as TasksTable } from "./components/workflow/TasksTable";
 export { default as SubmissionForm } from "./components/template/SubmissionForm";
 export { default as TemplateCard } from "./components/template/TemplateCard";
 export { default as SubmittedMessagesList } from "./components/workflow/SubmittedMessagesList";
+export { default as WorkflowsErrorBoundary } from "./components/workflow/WorkflowsErrorBoundary";
+export { default as WorkflowListFilterDrawer } from "./components/workflow/WorkflowListFilterDrawer";
+export { default as WorkflowsNavbar } from "./components/workflow/WorkflowsNavbar";
 export * from "./types";
